The north
This is a little project of mine from my university. It's silly but I wanted to do something different.

[ QUICK DISCLAIMER: The theme and subject have nothing to do with my ideals or form of living. I'm not implying any communist ideas ]

So everybody knows about the success of the Apollo 11 and the first men to land on the moon. But what about the other side of the competition, what about the USA cold wars enemy, the CCCP? Have they achieved something as well?
The mankind went to the moon on the Saturn V rocket, 3 months after the Russians tried to do so on the R1 rocket.
I planned to give you an insight into all this and let you think for yourself. Had the Americans really won the space race?